requirements.txt

pyyaml
loguru

gunicorn
flask
requests

